+++ src/build_options.rs
@@ -523,8 +523,7 @@ pub fn find_interpreter(
                     env::set_var("PYTHON_SYS_EXECUTABLE", &host_python.executable);
 
                     let sysconfig_path = find_sysconfigdata(cross_lib_dir.as_ref(), target)?;
-                    let sysconfig_data =
-                        parse_sysconfigdata(&host_python.executable, sysconfig_path)?;
+                    let sysconfig_data = parse_sysconfigdata(host_python, sysconfig_path)?;
                     let major = sysconfig_data
                         .get("version_major")
                         .context("version_major is not defined")?