+++ test-crates/pyo3-pure/check_installed/check_installed.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 import os
+import subprocess
 
 import pyo3_pure
 
@@ -9,5 +10,9 @@ assert pyo3_pure.DummyClass.get_42() == 42
 install_path = os.path.join(os.path.dirname(pyo3_pure.__file__))
 assert os.path.exists(os.path.join(install_path, "__init__.pyi"))
 assert os.path.exists(os.path.join(install_path, "py.typed"))
+
+# Check entrypoints (Unix only for now)
+if os.name != "nt":
+    assert subprocess.run(["get_42"]).returncode == 42
 
 print("SUCCESS")